644: Initial Compilation/port
Found out how to install Mega644 arduino stuff (See Makefile comments),
updated the Makefile to make it compatible to either 328 or 644 (or
both). It still default to 328, use make X4=1 to build the 644 version.
Also added a V=1 support to show the makefile progress when needed.

Added some updated pindef.h, altho seems the code doesn't use it that
much! At least the CD lines for the SD card are hard coded, so I'll have
to comb the code to find anything that is 'different' enough to require
fixing.

I haven't updated the bootloader, as I can't find the source for the
danII bootloader version, yet.

Signed-off-by: Michel Pollet <buserror@gmail.com>