Rationalisation of the pindef and usage
This commit replaces all the fixed hex constants for pins with the pin
names specified in pindefs.h
Also replaced some direct bit access with the macros that were already
defined for the same purpose.

This should not have any functional changes at all, it's just for sanity
sake as now we have two different pinouts for the two AVRs.

Signed-off-by: Michel Pollet <buserror@gmail.com>