x4: Fix SPI hangup
This is important, as it crashes the SPI (or, makes it look like it
crashes).
The SPI *will* drop of master mode, and this never complete a transfer
if the Slave Select pin goes low... if it was not used, and left
floating, it will randomly trigger, hanging the firmware.

This introduces a new SLAVE_SS define that will set the built-in pullup
on that pin.

Signed-off-by: Michel Pollet <buserror@gmail.com>