fwupdate: build: controvertial use of preprocessor for macros
So, this 'macros' the whole build phase, and will compile a 644p version
of the updater *IF* if finds a compiled 644p firmware.

Now it was not as simple as just building it, as the flash page size are
different, so I added a Makefile gizmo that passes teh .asm to the cc65
preprocessor first, allowing #ifdef etc, BEFORE passing it to the
assembler.

Well, it works, however it might not be the 'accepted' way of doing
apple II ASM... Here I wanted to have slightly different behaviour as we
now have 2 different page size, and I can't JUST do it by using a simple
macro, as in one case (128) it works, in the other case (256) it no
longer works as we're on 8 bits..

It works fine! Not entirely sure you aren't going to run around
screaming, but then again I'm unfamiliar with the assembler so...

Signed-off-by: Michel Pollet <buserror@gmail.com>