fwupdate: Fix the page overflow
The program could write too much before, this is now fixed by
calculating a max size in 'words' and comparing this in the loadAddress
subroutine. When the end is reached, a flag is set and return.

I've binary compared the file that is 'programmed' into the virtual
linapple card and the original fwimage.bin, and they are identical so
hopefully this is now all working.

Remaining todo will be to handle sizes much bigger, but that will
involves a much more complicated program, using the aux memory etc!

Signed-off-by: Michel Pollet <buserror@gmail.com>