fwupdate: Compile a 644 and a 328 fwupdate!
Now these is a big patch:
+ Makefile was parametrized so we can 'call' a macro to setup rules to
compile a 328P and a 644P version of the FWUpdate
+ The whole build now get files from Apple2Arduino/bin-*
+ the hex2bin utility was changed so it doesn't use the whole 'with
bootloader' file which as the whole flash, but uses the smaller
'application only' .hex file, and output a padded (to 256) binary image
instead.
+ The build system was change to pass the base dan2stk500.asm file via
the cc65 pre-processor with MCU specific #define, and generate an output
.asm for each MCU instead.

The build system will still work is the 644 isn't compiled, it'll just
build the 328p version only.

Signed-off-by: Michel Pollet <buserror@gmail.com>